refactor: combine all theme files
