--- packages/upswyng-web/src/theme/palette.ts
@@ -1,26 +1,0 @@
-const palette = {
-  common: { black: "#000", white: "#fff" },
-  background: { paper: "#262626", default: "#3a3a3a" },
-  primary: {
-    main: "#F05A28",
-    contrastText: "#000",
-  },
-  secondary: {
-    main: "#F4BD21",
-    contrastText: "#000",
-  },
-  error: {
-    main: "#BC2222",
-    contrastText: "#fff",
-  },
-  text: {
-    primary: "#fff",
-    secondary: "#CECECE",
-    disabled: "#fff",
-  },
-  action: {
-    active: "#fff",
-  },
-};
-
-export default palette;
