--- packages/upswyng-web/src/theme/typography.ts
@@ -1,27 +1,0 @@
-const typography = {
-  fontFamily: '"Open Sans","Helvetica","Arial",sans-serif',
-  fontSize: 16,
-  h1: {
-    fontSize: "1.5rem",
-    fontWeight: 400,
-  },
-  h2: {
-    fontSize: "1.25rem",
-    fontWeight: 600,
-  },
-  h3: {
-    fontSize: "1.25rem",
-    fontWeight: 300,
-  },
-  body1: {
-    fontSize: "1rem",
-    fontWeight: 300,
-  },
-  subtitle2: {
-    fontSize: "1rem",
-    fontWeight: 400,
-    margin: 0,
-  },
-};
-
-export default typography;
